# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v3/customer/{id}":{"get":{"tags":["customer-controller"],"operationId":"getBookById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}}}},"put":{"tags":["customer-controller"],"operationId":"updateCustomerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["customer-controller"],"operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/books/{bookId}":{"get":{"tags":["book-controller"],"operationId":"getBookById_1","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDTO"}}}}}},"put":{"tags":["book-controller"],"operationId":"updateById","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["book-controller"],"operationId":"deleteById","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"patch":{"tags":["book-controller"],"operationId":"updateBookPatchById","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v3/customer":{"get":{"tags":["customer-controller"],"operationId":"listCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerDTO"}}}}}}},"post":{"tags":["customer-controller"],"operationId":"handlePost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v2/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v2/auth/signin":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/books":{"get":{"tags":["book-controller"],"operationId":"listBooks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookDTO"}}}}}}},"post":{"tags":["book-controller"],"operationId":"handlePost_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/books/search":{"get":{"tags":["book-controller"],"operationId":"listBooksPageable","parameters":[{"name":"title","in":"query","required":false,"schema":{"type":"string"}},{"name":"bookGenre","in":"query","required":false,"schema":{"type":"string","enum":["FICTION","DRAMA","ROMANCE","SCIENCE_FICTION","FANTASY","HORROR","DOCUMENTARY","PSYCHOLOGICAL","POETRY","THRILLER","HISTORICAL_FICTION","BIOGRAPHY","MYSTERY","SCIENCE"]}},{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageBookDTO"}}}}}}}},"components":{"schemas":{"CustomerDTO":{"type":"object","properties":{"name":{"type":"string"},"id":{"type":"string","format":"uuid"},"version":{"type":"integer","format":"int32"},"email":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"updateDate":{"type":"string","format":"date-time"}}},"BookDTO":{"required":["price","title"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"version":{"type":"integer","format":"int32"},"isbn":{"type":"string"},"title":{"type":"string"},"price":{"type":"number"},"bookGenre":{"type":"string","enum":["FICTION","DRAMA","ROMANCE","SCIENCE_FICTION","FANTASY","HORROR","DOCUMENTARY","PSYCHOLOGICAL","POETRY","THRILLER","HISTORICAL_FICTION","BIOGRAPHY","MYSTERY","SCIENCE"]},"authors":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"publisher":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"updateDate":{"type":"string","format":"date-time"}}},"SignUpDTO":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"LoginDTO":{"type":"object","properties":{"usernameOrEmail":{"type":"string"},"password":{"type":"string"}}},"PageBookDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/BookDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}}}}}